+++ aflnet-replay.c
@@ -6,7 +6,7 @@
 
 #define server_wait_usecs 10000
 
-unsigned int* (*extract_response_codes)(unsigned char* buf, unsigned int buf_size, unsigned int* state_count_ref) = NULL;
+u64* (*extract_response_codes)(unsigned char* buf, unsigned int buf_size, unsigned int* state_count_ref) = NULL;
 
 /* Expected arguments:
 1. Path to the test case (e.g., crash-triggering input)
